src/main/java/com/example/elasticsearch/service/PersonService.java
// src/main/java/com/example/elasticsearch/service/PersonService.java

package com.example.elasticsearch.service;

import com.example.elasticsearch.model.Person;
import com.example.elasticsearch.repository.PersonRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class PersonService {

    @Autowired
    private PersonRepository personRepository;

    public Person savePerson(Person person) {
        return personRepository.save(person);
    }

    public List<Person> getAllPersons() {
        return (List<Person>) personRepository.findAll();
    }

    public Optional<Person> getPersonById(String id) {
        return personRepository.findById(id);
    }

    public void deletePerson(String id) {
        personRepository.deleteById(id);
    }
}
